Add cpu test
