+++ .ci/tritonbench/test-cpu.sh
@@ -2,4 +2,4 @@
 
 set -x
 
-python test/test_cpu/test_cpu.py
+python -m unittest test.test_cpu
