--- test/__init__.py
@@ -1,2 +1,0 @@
-from .test_cpu import load_cpu_tests
-from .test_gpu import load_gpu_tests
