--- test/test_gpu/test.py
@@ -1,3 +1,0 @@
-
-def load_gpu_tests():
-    pass
